+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/PlainSampleITs.java
@@ -42,8 +42,8 @@ public class PlainSampleITs extends AbstractIT {
       LibColumns.BOX_SEARCH, LibColumns.BOX_ALIAS, LibColumns.BOX_POSITION, LibColumns.DISCARDED, LibColumns.CREATION_DATE,
       LibColumns.WORKSTATION, LibColumns.THERMAL_CYCLER, LibColumns.PLATFORM, LibColumns.LIBRARY_TYPE, LibColumns.SELECTION,
       LibColumns.STRATEGY, LibColumns.INDEX_FAMILY, LibColumns.INDEX_1, LibColumns.INDEX_2, LibColumns.UMIS, LibColumns.KIT_DESCRIPTOR,
-      LibColumns.QC_PASSED, LibColumns.SIZE, LibColumns.CONCENTRATION, LibColumns.CONCENTRATION_UNITS, LibColumns.SPIKE_IN,
-      LibColumns.SPIKE_IN_DILUTION, LibColumns.SPIKE_IN_VOL);
+      LibColumns.KIT_LOT, LibColumns.QC_PASSED, LibColumns.SIZE, LibColumns.CONCENTRATION, LibColumns.CONCENTRATION_UNITS,
+      LibColumns.SPIKE_IN, LibColumns.SPIKE_IN_DILUTION, LibColumns.SPIKE_IN_VOL);
 
   private static final Set<String> libraryAliquotColumns = Sets.newHashSet(LibraryAliquotColumns.NAME, LibraryAliquotColumns.ALIAS,
       LibraryAliquotColumns.PARENT_ALIAS, LibraryAliquotColumns.BOX_SEARCH, LibraryAliquotColumns.BOX_ALIAS,