+++ RELEASE_NOTES.md
@@ -4,6 +4,11 @@ Changes:
 
   * Show included projects in Runs list
   * Allow archiving labs and institutes
+  * Added fields for tracking kit lot numbers
+    * Library kit
+    * Run sequencing kit
+    * Container clustering kit
+    * Container multiplexing ki
 
 # 1.1.0